extract ChannelJson
