+++ resources/assets/lib/stores/channel-store.ts
@@ -8,7 +8,8 @@ import { ChatNewConversationAdded } from 'actions/chat-new-conversation-added';
 import DispatcherAction from 'actions/dispatcher-action';
 import { dispatch, dispatchListener } from 'app-dispatcher';
 import ChatAPI from 'chat/chat-api';
-import { ChannelJson, ChannelType, GetUpdatesJson, MessageJson } from 'chat/chat-api-responses';
+import { GetUpdatesJson, MessageJson } from 'chat/chat-api-responses';
+import ChannelJson, { ChannelType } from 'interfaces/channel-json';
 import { groupBy, maxBy } from 'lodash';
 import { action, computed, makeObservable, observable, runInAction } from 'mobx';
 import Channel from 'models/chat/channel';
